reorg of item/components
