+++ common/src/main/scala/org/dka/rdbms/common/dao/TitleDao.scala
@@ -1,6 +1,6 @@
 package org.dka.rdbms.common.dao
 
-import org.dka.rdbms.common.model.ID
+import org.dka.rdbms.common.model.components.ID
 import org.dka.rdbms.common.model.item.Title
 
 /**
@@ -9,5 +9,3 @@ import org.dka.rdbms.common.model.item.Title
  * this interface is db agnostic and allows for easy unit testing since an database is not required
  */
 trait TitleDao extends CrudDao[Title, ID] {}
-
-
