--- common/src/main/scala/org/dka/rdbms/common/model/Item.scala
@@ -1,6 +1,0 @@
-package org.dka.rdbms.common.model
-
-trait Item[T] {
-  def value: T
-
-}
