+++ common/src/main/scala/org/dka/rdbms/common/model/components/PublishDate.scala
@@ -1,4 +1,7 @@
-package org.dka.rdbms.common.model
+package org.dka.rdbms.common.model.components
+
+import org.dka.rdbms.common.model.item.Item
+import org.dka.rdbms.common.model.validation.DateValidation
 
 import java.time.LocalDate
