+++ common/src/main/scala/org/dka/rdbms/common/model/item/Item.scala
@@ -1,0 +1,6 @@
+package org.dka.rdbms.common.model.item
+
+trait Item[T] {
+  def value: T
+
+}
