+++ common/src/main/scala/org/dka/rdbms/common/model/item/Publisher.scala
@@ -4,7 +4,8 @@ import cats.data.Validated._
 import cats.implicits._
 import io.circe._
 import org.dka.rdbms.common.model._
-import org.dka.rdbms.common.model.Validation._
+import org.dka.rdbms.common.model.components.{Address, City, CompanyName, ID, State, Zip}
+import org.dka.rdbms.common.model.validation.Validation._
 
 final case class Publisher(
   id: ID,
