+++ common/src/main/scala/org/dka/rdbms/common/model/validation/BigDecimalValidation.scala
@@ -1,24 +1,22 @@
-package org.dka.rdbms.common.model
+package org.dka.rdbms.common.model.validation
 
 import cats.data.Validated._
 import cats.implicits.catsSyntaxValidatedIdBinCompat0
 import io.circe._
+import org.dka.rdbms.common.model.item.Item
 
-import scala.util.{Try, Success, Failure}
+import scala.util.{Failure, Success, Try}
 
 trait BigDecimalValidation[T <: Item[BigDecimal]] extends Validation[String, BigDecimal, T] {
   import Validation._
 
-  def validate(string: String): ValidationErrorsOr[T] = {
-    Try { BigDecimal(string) }
-    match {
+  def validate(string: String): ValidationErrorsOr[T] =
+    Try(BigDecimal(string)) match {
       case Failure(t) => InvalidNumberException(fieldName, string, t).invalidNec
       case Success(date) => Valid(build(date))
     }
-  }
 
   def toJsonLine(item: T): (String, Json) = (fieldName, Json.fromString(item.value.toString))
-
 
   def fromOptionalJsonLine(c: HCursor): ValidationErrorsOr[Option[T]] = {
     val result = for {
