+++ slick/src/main/scala/org/dka/rdbms/slick/dao/AuthorDaoImpl.scala
@@ -1,8 +1,9 @@
 package org.dka.rdbms.slick.dao
 
 import org.dka.rdbms.common.dao.AuthorDao
+import org.dka.rdbms.common.model.components.{Address, City, FirstName, ID, LastName, Phone, State, Zip}
 import org.dka.rdbms.common.model.item.Author
-import org.dka.rdbms.common.model.{Address, City, FirstName, ID, LastName, Phone, State, Zip, item}
+import org.dka.rdbms.common.model.item
 import slick.jdbc.JdbcBackend.Database
 import slick.jdbc.PostgresProfile.api._
 import slick.lifted.TableQuery
