+++ slick/src/test/scala/org/dka/rdbms/slick/dao/AuthorDaoImplSpec.scala
@@ -2,8 +2,9 @@ package org.dka.rdbms.slick.dao
 
 import com.typesafe.scalalogging.Logger
 import org.dka.rdbms.TearDownException
-import org.dka.rdbms.common.model.{Address, City, FirstName, ID, LastName, Phone, State, Zip, item}
+import org.dka.rdbms.common.model.item
 import AuthorDaoImplSpec._
+import org.dka.rdbms.common.model.components.{Address, City, FirstName, ID, LastName, Phone, State, Zip}
 import org.dka.rdbms.common.model.item.Author
 import org.scalatest.funspec.AnyFunSpec
 import org.scalatest.matchers.should.Matchers
